server/requirements.txt
fastapi
uvicorn
python-socketio
bcrypt
PyJWT
firebase-admin
blind-watermark